% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{clustered_sample}
\alias{clustered_sample}
\title{Simulates clustered samples.}
\usage{
clustered_sample(sarea, nsamples, nparents, radius)
}
\arguments{
\item{sarea}{sf/sfc polygon where samples will be simulated.}

\item{nsamples}{Integer. Number of samples to simulate.}

\item{nparents}{Integer. Number of parents to simulate.}

\item{radius}{Numeric. Radius of the buffer for children simulation.}
}
\description{
Simulates clustered samples.
}
\details{
Simulates clustered samples by simulating a number of randomly sampled
parents, and then randomly simulate children within a buffer of the parents.
}
