% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{jittered_sample}
\alias{jittered_sample}
\title{Simulates regular sample jittered by an amount of noise}
\usage{
jittered_sample(x, size, amount, ...)
}
\arguments{
\item{x}{sf/sfc polygon where samples will be simulated.}

\item{size}{Number of samples to create.}

\item{amount}{Amount of jitter to apply.}

\item{...}{Additional arguments passed to \code{terra::spatSample()}.}
}
\description{
Simulates regular sample jittered by an amount of noise ~ U(-amount, amount).
}
