% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{jitterreg_sample}
\alias{jitterreg_sample}
\title{Simulates regular samples jittered by an amount of noise.}
\usage{
jitterreg_sample(sarea, nsamples, amount)
}
\arguments{
\item{sarea}{sf/sfc polygon where samples will be simulated.}

\item{nsamples}{Integer. Number of samples to simulate.}

\item{amount}{Numeric. Amount of jitter to apply.}
}
\description{
Simulates regular samples jittered by an amount of noise.
}
\details{
Simulates regular samples jittered by an amount of noise ~ U(-amount, amount).
}
