% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proxy.R
\name{make_proxy}
\alias{make_proxy}
\alias{add_proxy}
\title{Generate spatial proxies}
\usage{
make_proxy(rast_grid, method = proxy_coordinates())

add_proxy(rast_grid, method = proxy_coordinates())
}
\arguments{
\item{rast_grid}{A \code{SpatRaster}}

\item{method}{A proxy method created by \verb{proxy_*()} functions}
}
\value{
A \code{SpatRaster} object

\code{SpatRaster}
}
\description{
The \code{make_proxy()} function generates spatial proxies for a given raster object.
It generates spatial proxies such as:
\itemize{
\item coordinates
\item Euclidean Distance Fields (EDF)
\item Oblique Geographic Coordinates (OGC)
}
}
\examples{
rast_grid = terra::rast(
  ncols = 300, nrows = 100,
  xmin = 0, xmax = 300,
  ymin = 0, ymax = 100
)

proxy_coords = make_proxy(rast_grid, proxy_coordinates())
proxy_edf = make_proxy(rast_grid, proxy_edf())
proxy_ogc = make_proxy(rast_grid, proxy_ogc(n = 5))

terra::plot(proxy_coords)
terra::plot(proxy_edf)
terra::plot(proxy_ogc)
}
\references{
Behrens, T., Schmidt, K., Viscarra Rossel, R. A., Gries, P., Scholten, T., & MacMillan, R. A. (2018). Spatial modelling with Euclidean distance fields and machine learning. European journal of soil science, 69(5), 757-770.

Møller, A. B., Beucher, A. M., Pouladi, N., & Greve, M. H. (2020). Oblique geographic coordinates as covariates for digital soil mapping. Soil, 6(2), 269-289.
}
