% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proxy.R
\name{proxy_coordinates}
\alias{proxy_coordinates}
\title{Coordinate proxy factory}
\usage{
proxy_coordinates()
}
\value{
A function that accepts \code{rast_grid} and returns a \code{SpatRaster}
}
\description{
Returns a function that generates coordinate layers (X, Y) for a raster.
}
\examples{
rast_grid = terra::rast(
  ncols = 300, nrows = 100,
  xmin = 0, xmax = 300,
  ymin = 0, ymax = 100
)

make_proxy(rast_grid, proxy_coordinates())
}
