% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proxy.R
\name{proxy_ogc}
\alias{proxy_ogc}
\title{Oblique Geographic Coordinates proxy factory}
\usage{
proxy_ogc(n = 5)
}
\arguments{
\item{n}{Number of angles (default: 5)}
}
\value{
A function that accepts \code{rast_grid} and returns a \code{SpatRaster}
}
\description{
Returns a function that generates Oblique Geographic Coordinates (OGC) as covariates.
}
\examples{
rast_grid = terra::rast(
  ncols = 300, nrows = 100,
  xmin = 0, xmax = 300,
  ymin = 0, ymax = 100
)

make_proxy(rast_grid, proxy_ogc(n = 5))
}
\references{
Møller, A. B., Beuchert, A. M., Pouladi, N., & Greve, M. H. (2020).
Oblique geographic coordinates as covariates for digital soil mapping.
Soil, 6(2), 269-289.
}
