% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Evaluate quietly}
\usage{
quiet(x)
}
\arguments{
\item{x}{A function to evaluate.}
}
\description{
Evaluate quietly
}
\details{
Function to hide undesired prints. Copied from:
https://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html
}
\examples{
# quiet(print("Nothing is shown"))
}
