% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{sam_field}
\alias{sam_field}
\title{Sample spatial field}
\usage{
sam_field(x, size, method = sample_random())
}
\arguments{
\item{x}{A raster object (\code{SpatRaster}).}

\item{size}{Number of samples to create.}

\item{method}{Sampling method (function created by \code{\link[=sample_random]{sample_random()}},
\code{\link[=sample_jittered]{sample_jittered()}}, or \code{\link[=sample_clustered]{sample_clustered()}}).}
}
\value{
An \code{sf} object with sampled points
}
\description{
Create a sample of a spatial field.
}
\examples{
rast_grid = terra::rast(
  ncols = 300, nrows = 100,
  xmin = 0, xmax = 300,
  ymin = 0, ymax = 100
)

sam_field(
  rast_grid,
  100,
  method = sample_jittered(amount = 5)
)

}
