% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{sample_clustered}
\alias{sample_clustered}
\title{Clustered sampling method}
\usage{
sample_clustered(nclusters, radius, ...)
}
\arguments{
\item{nclusters}{Number of clusters}

\item{radius}{Cluster radius (in map units)}

\item{...}{Additional arguments passed to \code{terra::spatSample()}}
}
\value{
A function that accepts \code{x} (SpatRaster) and \code{size} and returns an \code{sf} object
}
\description{
Returns a function that performs clustered sampling on a spatial field.
Cluster centers are randomly distributed, then samples are drawn within
a buffer radius around each cluster.
}
\examples{
rast_grid = terra::rast(
  ncols = 300, nrows = 100,
  xmin = 0, xmax = 300,
  ymin = 0, ymax = 100
)

sam_field(rast_grid, 200, method = sample_clustered(nclusters = 5, radius = 10))
}
