% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{sample_jittered}
\alias{sample_jittered}
\title{Jittered sampling method}
\usage{
sample_jittered(amount, ...)
}
\arguments{
\item{amount}{Jitter amount in map units}

\item{...}{Additional arguments passed to \code{terra::spatSample()}}
}
\value{
A function that accepts \code{x} (SpatRaster) and \code{size} and returns an \code{sf} object
}
\description{
Returns a function that performs jittered sampling on a spatial field.
Regular points are generated first, then jittered by a random amount.
}
\examples{
rast_grid = terra::rast(
  ncols = 300, nrows = 100,
  xmin = 0, xmax = 300,
  ymin = 0, ymax = 100
)

sam_field(rast_grid, 100, method = sample_jittered(amount = 5))
}
