% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simsam.R
\name{simsam}
\alias{simsam}
\title{Simulate Randomly Generated Correlated Data}
\usage{
simsam(orig_dat, nrows, your_choice_of_NA = NULL)
}
\arguments{
\item{orig_dat}{A data set the user wishes to generate a random, correlated
data set for.}

\item{nrows}{The number of rows to make the output data set.}

\item{your_choice_of_NA}{A vector of characters that denote missing values
(e.g., "N/A", 999).}
}
\value{
Returns a data set of randomly generated data corelated with the
riginal data set.
}
\description{
Simulate n rows of a randomly generated data set from an existing data set.
General structure (column means and correlations) are preserved.
}
\examples{
library(dplyr)
(sim_dat <- simsam(mtcars, 1000)[[1]] \%>\% tbl_df())
colMeans(sim_dat); colMeans(mtcars)
cor(sim_dat); cor(mtcars)
}

