% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_covariates.R
\name{simulate_random}
\alias{simulate_random}
\title{Simulate uncorrelated random fields}
\usage{
simulate_random(mean = 0, sd = 1, seed = NULL)
}
\arguments{
\item{mean}{Mean of the random values (default: 0)}

\item{sd}{Standard deviation of the random values (default: 1)}

\item{seed}{Optional random seed for reproducibility}
}
\value{
A function that accepts \code{rast_grid} and \code{n} and returns a \code{SpatRaster}
}
\description{
Returns a function that simulates uncorrelated random fields (white noise).
}
\examples{
rast_grid = terra::rast(ncols = 300, nrows = 100,
  xmin = 0, xmax = 300, ymin = 0, ymax = 100)
sim_fn = simulate_random()
sf1 = sim_covariates(rast_grid, n = 4, method = sim_fn)
terra::plot(sf1)
}
