% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{clustered_sample}
\alias{clustered_sample}
\title{Simulates clustered samples}
\usage{
clustered_sample(x, size, nclusters, radius, ...)
}
\arguments{
\item{x}{sf/sfc polygon where samples will be simulated.}

\item{size}{Number of samples to create.}

\item{nclusters}{Number of clusters to simulate.}

\item{radius}{Radius of the buffer for intra-cluster simulation.}

\item{...}{Additional arguments passed to \code{terra::spatSample()}.}
}
\description{
Simulates clustered samples by simulating a number of randomly sampled
clusters, and then randomly simulate points within a buffer of the clusters.
}
