% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{jitterreg_sample}
\alias{jitterreg_sample}
\title{Simulates regular sample jittered by an amount of noise}
\usage{
jitterreg_sample(x, size, amount)
}
\arguments{
\item{x}{sf/sfc polygon where samples will be simulated.}

\item{size}{Number of samples to create.}

\item{amount}{Amount of jitter to apply.}
}
\description{
Simulates regular sample jittered by an amount of noise ~ U(-amount, amount).
}
