% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{sam_field}
\alias{sam_field}
\title{Sample spatial field}
\usage{
sam_field(x, size, type = "", type_opts = NULL, ...)
}
\arguments{
\item{x}{A raster object.}

\item{size}{Number of samples to create.}

\item{type}{Expected spatial distribution of the samples. Possible values are:
\code{"jittered"}, \code{"random"}, \code{"clustered"}.}

\item{type_opts}{A list of options specific to the selected \code{type}.
When \code{type = "jittered"}, the \code{amount} option specifies the amount of jitter to apply to the samples (in map units).
When \code{type = "clustered"}, the \code{nclusters} option specifies the number of clusters to simulate, and the \code{radius} option specifies the radius of the buffer around each cluster.}

\item{...}{Additional arguments passed to \code{terra::spatSample()}.}
}
\description{
Create a sample of a spatial field.
}
\examples{
rast_grid = terra::rast(ncols = 300, nrows = 100, xmin = 0, xmax = 300, ymin = 0, ymax = 100)
sam_field(rast_grid, 100, "jittered", type_opts = list(amount = 5))
}
