% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{sam_field}
\alias{sam_field}
\title{Sample spatial field}
\usage{
sam_field(x, size, type, value, nclusters, ...)
}
\arguments{
\item{x}{Any object that can be coerced to an sf object.}

\item{size}{Number of samples to create.}

\item{type}{Expected spatial distribution of the samples. Possible values are:
\code{"jittered"}, \code{"random"}, \code{"clustered"}.}

\item{value}{Amount of jitter to apply to the samples (in map units; only used when \code{type = "jittered"})
or radius of the buffer around each cluster (only used when \code{type = "clustered"}).}

\item{nclusters}{Number of clusters to simulate. Only used when \code{type = "clustered"}.}

\item{...}{Additional arguments passed to \code{sf::st_sample()} when \code{type = "random"}.}
}
\description{
Create a sample of a spatial field.
}
\examples{
rast_grid = terra::rast(ncols = 300, nrows = 100, xmin = 0, xmax = 300, ymin = 0, ymax = 100)
sam_field(rast_grid, 100, "jittered", 5)
vect_grid = sf::st_as_sf(terra::as.polygons(rast_grid))
sam_field(vect_grid, 100, "random")
}
