% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_field.R
\name{sam_field}
\alias{sam_field}
\title{Sample simulation}
\usage{
sam_field(sarea, nsamples, dsamples)
}
\arguments{
\item{nsamples}{Integer. Number of samples to simulate.}

\item{dsamples}{Character. Spatial distribution of the samples. 5 are
possible: sregular, wregular, random, wclust, sclust.}
}
\description{
Sample simulation
}
\details{
Simulates a series of sampling points for simulation problem 1.
}
