% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_covariates.R
\name{sim_covariates}
\alias{sim_covariates}
\title{Simulate a raster stack of covariates based on a variogram model}
\usage{
sim_covariates(rast_grid, vgm = NULL, n = 6, ...)
}
\arguments{
\item{rast_grid}{A raster object with the desired dimensions}

\item{vgm}{A variogram model object}

\item{n}{The number of covariates to simulate}

\item{...}{Additional arguments to be passed to the \code{gstat::vgm()} function; \code{beta}, \code{nmax}, \code{indicators} (see Examples)}
}
\value{
A raster object with the simulated covariates
}
\description{
Simulate a raster stack of covariates based on a variogram model
}
\examples{
rast_grid = terra::rast(ncols = 300, nrows = 100,
  xmin = 0, xmax = 300, ymin = 0, ymax = 100)
sf1 = sim_covariates(rast_grid, range = 25,
  vgm = gstat::vgm(model = "Exp", psill = 1, range = 10), n = 4)
sf2 = sim_covariates(rast_grid, range = 50, n = 4)
sf3 = sim_covariates(rast_grid, range = 25,
  vgm = gstat::vgm(model = "Nug", psill = 2, range = 0), n = 4)
sf4 = sim_covariates(rast_grid, range = 25,
  vgm = gstat::vgm(model = "Exp", psill = 10, range = 100), n = 1,
  beta = 30, indicators = TRUE)
terra::plot(sf1)
terra::plot(sf2)
terra::plot(sf3)
terra::plot(sf4)
}
