% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsam-package.R
\docType{package}
\name{simsam-package}
\alias{simsam}
\alias{simsam-package}
\title{simsam: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

Other contributors:
\itemize{
  \item Carles Milà [contributor]
}

}
\keyword{internal}
